Refactored UITests code
